requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
basepage==1.7.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
content==0.0.1
cryptography==41.0.4
decorator==5.1.1
Deprecated==1.2.14
driver==0.1.0
exceptiongroup==1.1.3
flake8==6.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
locator==0.1.0
lxml==4.9.3
mccabe==0.7.0
outcome==1.2.0
packaging==23.1
pillow==10.2.0
pluggy==1.3.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==1.10.12
pyflakes==3.1.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3.post1
reportlab==4.0.9
requests==2.31.0
selenium==4.11.2
self==2020.12.3
slack_sdk==3.26.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
testrail_api==1.13.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.8.0
urllib3==2.0.4
vonage==2.5.5
vonage-jwt==1.0.0
webdriver-manager==4.0.0
wrapt==1.15.0
wsproto==1.2.0
